moved seed identity to it's own class
